+++ src/aceinna/devices/configs/rtk_predefine.py
@@ -1,11 +1,12 @@
 """
-predefined params for openrtk
+predefined params for openrtk 
 """
 
 
 def get_rtk_products():
     return {
-        'RTK330L': ['RTK_INS']
+        'RTK330L': ['RTK_INS'],
+        'RTK350L': ['RTK_INS']
     }
 
 
@@ -19,7 +20,8 @@ def get_app_names():
 
 def get_configuratin_file_mapping():
     return {
-        'RTK330L': 'RTK330L.json'
+        'RTK330L': 'RTK330L.json',
+        'RTK350L': 'RTK350LA.json'
     }