+++ src/aceinna/devices/parsers/ins401_field_parser.py
@@ -53,25 +53,25 @@ def do_decode_value(data_type, data, conf):
         return unpack_value
     elif data_type == 'uint16':
         try:
-            pack_item = struct.pack('2B', *data)
+            pack_item = struct.pack('2B', *data[0:2])
         except:  # pylint: disable=bare-except
             return False
         return struct.unpack('<H', pack_item)[0]
     elif data_type == 'int16':
         try:
-            pack_item = struct.pack('2B', *data)
+            pack_item = struct.pack('2B', *data[0:2])
         except:  # pylint: disable=bare-except
             return False
         return struct.unpack('<h', pack_item)[0]
     elif data_type == 'uint8':
         try:
-            pack_item = struct.pack('1B', *data)
+            pack_item = struct.pack('1B', data[0])
         except:  # pylint: disable=bare-except
             return False
         return struct.unpack('<B', pack_item)[0]
     elif data_type == 'int8':
         try:
-            pack_item = struct.pack('1B', *data)
+            pack_item = struct.pack('1B', data[0])
         except:  # pylint: disable=bare-except
             return False
         return struct.unpack('<b', pack_item)[0]